Fix size of structure member

+++ pulseaudio/lib_pulseaudio.py
@@ -215,7 +215,7 @@ class PA_SAMPLE_SPEC(Structure):
   _fields_ = [
     ("format", c_int), # FIXME check this
     ("rate", c_uint32),
-    ("channels", c_uint32)
+    ("channels", c_uint8)
   ]
 
 class PA_CHANNEL_MAP(Structure):
